api/requirements.txt
dataclasses==0.7
Django==2.2.7
django-cors-headers==3.2.0
djangorestframework==3.10.3
djongo==1.2.38
pymongo==3.9.0
pytz==2019.3
sqlparse==0.2.4
